+++ out/game-data.js
@@ -131,6 +131,14 @@ exports.Pokemons = {
                 { name: 'Feraligatr', sprite: 'feraligatr', spriteSize: 32, candyCost: 25 },
             ],
         },
+        {
+            name: 'Togepi',
+            forms: [
+                { name: 'Togepi', sprite: 'togepi', spriteSize: 32, candyCost: 0 },
+                { name: 'Togetic', sprite: 'togetic', spriteSize: 32, candyCost: 10 },
+                { name: 'Togekiss', sprite: 'togekiss', spriteSize: 32, candyCost: 25, requiredItem: 'shiny_stone', requiredFriendship: 200 },
+            ],
+        },
     ],
     'generation 3': [
         { name: 'Treecko', forms: [{ name: 'Treecko', sprite: 'treecko', spriteSize: 32, candyCost: 0 }, { name: 'Grovyle', sprite: 'grovyle', spriteSize: 32, candyCost: 10 }, { name: 'Sceptile', sprite: 'sceptile', spriteSize: 32, candyCost: 25 }] },
@@ -141,6 +149,8 @@ exports.Pokemons = {
         { name: 'Turtwig', forms: [{ name: 'Turtwig', sprite: 'turtwig', spriteSize: 32, candyCost: 0 }, { name: 'Grotle', sprite: 'grotle', spriteSize: 32, candyCost: 10 }, { name: 'Torterra', sprite: 'torterra', spriteSize: 48, candyCost: 25 }] },
         { name: 'Chimchar', forms: [{ name: 'Chimchar', sprite: 'chimchar', spriteSize: 32, candyCost: 0 }, { name: 'Monferno', sprite: 'monferno', spriteSize: 32, candyCost: 10 }, { name: 'Infernape', sprite: 'infernape', spriteSize: 32, candyCost: 25 }] },
         { name: 'Piplup', forms: [{ name: 'Piplup', sprite: 'piplup', spriteSize: 32, candyCost: 0 }, { name: 'Prinplup', sprite: 'prinplup', spriteSize: 32, candyCost: 10 }, { name: 'Empoleon', sprite: 'empoleon', spriteSize: 32, candyCost: 25 }] },
+        { name: 'Shinx', forms: [{ name: 'Shinx', sprite: 'shinx', spriteSize: 32, candyCost: 0 }, { name: 'Luxio', sprite: 'luxio', spriteSize: 32, candyCost: 10 }, { name: 'Luxray', sprite: 'luxray', spriteSize: 32, candyCost: 25 }] },
+        { name: 'Riolu', forms: [{ name: 'Riolu', sprite: 'riolu', spriteSize: 32, candyCost: 0 }, { name: 'Lucario', sprite: 'lucario', spriteSize: 32, candyCost: 10, requiredFriendship: 220 }] },
     ],
     'generation 5': [
         { name: 'Snivy', forms: [{ name: 'Snivy', sprite: 'snivy', spriteSize: 32, candyCost: 0 }, { name: 'Servine', sprite: 'servine', spriteSize: 32, candyCost: 10 }, { name: 'Serperior', sprite: 'serperior', spriteSize: 32, candyCost: 25 }] },